Fix empty value overrides
